Move to skaffold package
